Pure renaming PR: ampelement.css -> ampshared.css
