+++ build-system/tasks/presubmit-checks.js
@@ -56,7 +56,7 @@ const forbiddenTerms = {
       'build-system/app-index/boilerplate.js',
       'build-system/tasks/extension-generator/index.js',
       'css/ampdoc.css',
-      'css/ampelement.css',
+      'css/ampshared.css',
       'extensions/amp-pinterest/0.1/amp-pinterest.css',
       'extensions/amp-pinterest/0.1/follow-button.js',
       'extensions/amp-pinterest/0.1/pin-widget.js',
@@ -69,7 +69,7 @@ const forbiddenTerms = {
       'build-system/tasks/extension-generator/index.js',
       'build-system/tasks/create-golden-css/css/main.css',
       'css/ampdoc.css',
-      'css/ampelement.css',
+      'css/ampshared.css',
     ],
   },
   'describe\\.only': '',
