+++ css/ampshared.css
