+++ src/inabox/amp-inabox-lite.js
@@ -26,7 +26,7 @@ import {
 import {Services} from '../services';
 import {adopt, installBuiltins, installRuntimeServices} from '../runtime';
 import {cssText as ampDocCss} from '../../build/ampdoc.css';
-import {cssText as ampElementCss} from '../../build/ampelement.css';
+import {cssText as ampElementCss} from '../../build/ampshared.css';
 import {fontStylesheetTimeout} from '../font-stylesheet-timeout';
 import {getMode} from '../mode';
 import {installDocService} from '../service/ampdoc-impl';
