+++ src/runtime.js
@@ -33,7 +33,7 @@ import {
 import {Services} from './services';
 import {VisibilityState} from './visibility-state';
 import {cssText as ampDocCss} from '../build/ampdoc.css';
-import {cssText as ampElementCss} from '../build/ampelement.css';
+import {cssText as ampElementCss} from '../build/ampshared.css';
 import {
   childElementsByTag,
   isConnectedNode,
